refactor: split sandbox snapshot module
